gemspec: depend on sinatra-contrib version 2
The package, version 1.4.7, in Debian currently fails to build from
source. If left unfixed, this will lead to automatic removal, which in
turn will remove schleuder as well. The package was maintained by Lunar,
probably, I'll take over this package. If so, I'll first upload the
current version 2.0.6; hence, requiring it here as well.