Drop pin_keys (originally introduced to support v2 to v3 migration)
Remove obsolete code, which looked for subscriptions without an
associated key, and tried to find a distinctly matching key. Originally
this was implemented to help with a shortcoming of code which handled v2
to v3 migration. As the support to migrate v2 lists to v3 is dropped in
v4, remove the pinning code as well.

Closes #411

(cherry picked from commit 0eb91fddb3b100711a3625de4dc69b12fa3d8fda)